updated register page
